+++ .storybook/config.js
@@ -1,7 +1,10 @@
-import { configure } from '@storybook/react';
+import { configure } from "@storybook/react";
+import { addDecorator } from "@storybook/react";
+import { withKnobs } from "@storybook/addon-knobs/react";
 
+addDecorator(withKnobs);
 // automatically import all files ending in *.stories.js
-const req = require.context('../stories', true, /.stories.js?x$/);
+const req = require.context("../stories", true, /.stories.js?x$/);
 function loadStories() {
   req.keys().forEach(filename => req(filename));
 }