+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.0.36 (2018-05-29)
+
+* new prop **CustomTd** React.ComponentType<CalendarDayProps> - shoudl be used as custom Day component in Calendar
+
+# 0.0.35 (2018-05-17)
+
+* new props **customClasses** { [string]: Date[]} - the class (css) will be applied to the dates that are in the array
+
 # 0.0.34 (2018-03-23)
 
 * **styled-components** are now `peerDependency`